+++ src/api-http-presenter/index.ts
@@ -2,10 +2,12 @@ import defaultMiddlewares from './middlewares/defaultMiddlewares';
 import NotesAPIRoutes from './routes/NotesAPIRoutes';
 import FlashcardsAPIRoutes from './routes/FlashcardsAPIRoutes';
 import StatsAPIRoutes from './routes/LevelingAPIRoutes';
+import FeatsAPIRoutes from './routes/FeatsAPIRoutes';
 
 export default [
   ...defaultMiddlewares,
   ...NotesAPIRoutes,
   ...FlashcardsAPIRoutes,
+  ...FeatsAPIRoutes,
   ...StatsAPIRoutes
 ]
