+++ src/repositories/FeatsRepo.ts
@@ -1,6 +1,6 @@
-import { IFeats } from "../../features/interfaces/interfaces";
-import { RepositoryError } from "../../util/errors/RepositoryError"
-import { NoSQLRepository } from "../RepoResultHandler"
+import { IFeats } from "./../features/interfaces/interfaces";
+import { RepositoryError } from "./../util/errors/RepositoryError"
+import { NoSQLRepository } from "./RepoResultHandler"
 import { ObjectId } from 'mongodb'
 
 const dbName = 'leveling';
@@ -15,7 +15,7 @@ class FeatsRepo extends NoSQLRepository<IFeats>{
   }
 
   async findOneFeat(identifier: uniqueIdentifier) {
-    const searchParams = this.searchParams(identifier);
+    const searchParams = buildSearchString(identifier);
 
     const record = await this.collection().findOne(searchParams);
     return { record };
@@ -76,13 +76,13 @@ class FeatsRepo extends NoSQLRepository<IFeats>{
     if (invalidProperty)
       throw new RepositoryError('Invalid property issued: '+ invalidProperty);
       
-    const searchParams = this.searchParams(identifier);
+    const searchParams = buildSearchString(identifier);
 
     await this.collection().findOneAndUpdate(searchParams, {$set: properties});
   }
 
   async proceedAcceptanceLevel(identifier: uniqueIdentifier, stage: 'reviewed'|'ready') {
-    const searchParams = this.searchParams(identifier);
+    const searchParams = buildSearchString(identifier);
     await this.collection().findOneAndUpdate(searchParams, {
       $set:{"acceptance.stage":stage},
       $push: {"acceptance.date": new Date()}
@@ -90,23 +90,23 @@ class FeatsRepo extends NoSQLRepository<IFeats>{
   }
 
   async deleteOneFeat(identifier: uniqueIdentifier) {
-    const searchParams = this.searchParams(identifier);    
+    const searchParams = buildSearchString(identifier);    
     await this.collection().findOneAndDelete(searchParams);
   }
 
-  private searchParams(identifier: uniqueIdentifier) {
-    let result:{title: string}|{_id: ObjectId};
-    const { title, _id } = identifier;
-
-    if (title)
-      result = {title};
-    else if(_id)
-      result = {_id: new ObjectId(_id)};
-    else
-      throw new Error('An identifier must be specified.');
-
-    return result;
-  }
 }
 
+export function buildSearchString(identifier: uniqueIdentifier) {
+  let result:{title: string}|{_id: ObjectId};
+  const { title, _id } = identifier;
+
+  if (title)
+    result = {title};
+  else if(_id)
+    result = {_id: new ObjectId(_id)};
+  else
+    throw new Error('An identifier must be specified.');
+
+  return result;
+}
 export default new FeatsRepo(dbName, collectionName);
